Move AwtWindow, SwingWindow and SwingDialog into the androidx.compose.ui.awt package
