+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/window/Dialog.desktop.kt
@@ -18,16 +18,14 @@ package androidx.compose.ui.window
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
-import androidx.compose.runtime.currentCompositionLocalContext
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.awt.ComposeDialog
+import androidx.compose.ui.awt.SwingDialog
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.input.key.KeyEvent
-import androidx.compose.ui.platform.LocalLayoutDirection
-import androidx.compose.ui.scene.LocalComposeSceneContext
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.util.ComponentUpdater
@@ -118,12 +116,12 @@ fun DialogWindow(
 }
 
 /**
- * Composes platform dialog in the current composition. When Dialog enters the composition,
- * a new platform dialog will be created and receives the focus. When Dialog leaves the
- * composition, dialog will be disposed and closed.
+ * Composes platform dialog in the current composition. When [DialogWindow] enters the composition,
+ * a new platform dialog will be created and receive focus. When [DialogWindow] leaves the
+ * composition, the dialog will be disposed and closed.
  *
- * Dialog is a modal window. It means it blocks the parent [Window] / [DialogWindow] in which composition
- * context it was created.
+ * Dialog is a modal window. It means it blocks the parent [Window] / [DialogWindow] in whose
+ * composition context it was created.
  *
  * Usage:
  * ```
@@ -131,7 +129,7 @@ fun DialogWindow(
  * fun main() = application {
  *     var isDialogOpen by remember { mutableStateOf(true) }
  *     if (isDialogOpen) {
- *         Dialog(onCloseRequest = { isDialogOpen = false }) {}
+ *         DialogWindow(onCloseRequest = { isDialogOpen = false }) {}
  *     }
  * }
  * ```
@@ -146,34 +144,35 @@ fun DialogWindow(
  * the native dialog will update its corresponding properties.
  * If [DialogState.position] is not [WindowPosition.isSpecified], then after the first show on the
  * screen [DialogState.position] will be set to the absolute values.
- * @param visible Is [DialogWindow] visible to user.
+ * @param visible Whether the dialog is visible to the user.
  * If `false`:
  * - internal state of [DialogWindow] is preserved and will be restored next time the dialog
  * will be visible;
  * - native resources will not be released. They will be released only when [DialogWindow]
  * will leave the composition.
- * @param title Title in the titlebar of the dialog
- * @param icon Icon in the titlebar of the window (for platforms which support this).
+ * @param title Title in the title bar of the dialog
+ * @param icon Icon in the title bar of the dialog (for platforms that support this).
  * On macOs individual windows can't have a separate icon. To change the icon in the Dock,
  * set it via `iconFile` in build.gradle
  * (https://github.com/JetBrains/compose-jb/tree/master/tutorials/Native_distributions_and_local_execution#platform-specific-options)
- * @param undecorated Disables or enables decorations for this window.
- * @param transparent Disables or enables window transparency. Transparency should be set
- * only if window is undecorated, otherwise an exception will be thrown.
- * @param resizable Can dialog be resized by the user (application still can resize the dialog
- * changing [state])
- * @param enabled Can dialog react to input events
- * @param focusable Can dialog receive focus
- * @param alwaysOnTop Should the dialog always be on top of another windows and dialogs
+ * @param undecorated Disables or enables decorations for this dialog.
+ * @param transparent Disables or enables window transparency. Transparency may be set only if the
+ * dialog is undecorated, otherwise an exception will be thrown.
+ * @param resizable Whether the dialog can be resized by the user (application still can resize the
+ * dialog by changing [state]).
+ * @param enabled Whether the dialog reacts to input events.
+ * @param focusable Whether the dialog can receive focus.
+ * @param alwaysOnTop whether the dialog will always be on top of other windows and dialogs in the
+ * application.
  * @param onPreviewKeyEvent This callback is invoked when the user interacts with the hardware
  * keyboard. It gives ancestors of a focused component the chance to intercept a [KeyEvent].
  * Return true to stop propagation of this event. If you return false, the key event will be
  * sent to this [onPreviewKeyEvent]'s child. If none of the children consume the event,
- * it will be sent back up to the root using the onKeyEvent callback.
+ * it will be sent back up to the root using the [onKeyEvent] callback.
  * @param onKeyEvent This callback is invoked when the user interacts with the hardware
  * keyboard. While implementing this callback, return true to stop propagation of this event.
  * If you return false, the key event will be sent to this [onKeyEvent]'s parent.
- * @param content content of the dialog
+ * @param content Composable content of the dialog.
  */
 @Composable
 fun DialogWindow(
@@ -211,12 +210,12 @@ fun DialogWindow(
 }
 
 /**
- * Composes platform dialog in the current composition. When Dialog enters the composition,
- * a new platform dialog will be created and receives the focus. When Dialog leaves the
- * composition, dialog will be disposed and closed.
+ * Composes platform dialog in the current composition. When [DialogWindow] enters the composition,
+ * a new platform dialog will be created and receive focus. When [DialogWindow] leaves the
+ * composition, the dialog will be disposed and closed.
  *
- * Dialog is a modal window. It means it blocks the parent [Window] / [DialogWindow] in which composition
- * context it was created.
+ * Dialog is a modal window. It means it blocks the parent [Window] / [DialogWindow] in whose
+ * composition context it was created.
  *
  * Usage:
  * ```
@@ -224,7 +223,7 @@ fun DialogWindow(
  * fun main() = application {
  *     var isDialogOpen by remember { mutableStateOf(true) }
  *     if (isDialogOpen) {
- *         Dialog(onCloseRequest = { isDialogOpen = false }) {}
+ *         DialogWindow(onCloseRequest = { isDialogOpen = false }) {}
  *     }
  * }
  * ```
@@ -239,34 +238,35 @@ fun DialogWindow(
  * the native dialog will update its corresponding properties.
  * If [DialogState.position] is not [WindowPosition.isSpecified], then after the first show on the
  * screen [DialogState.position] will be set to the absolute values.
- * @param visible Is [DialogWindow] visible to user.
+ * @param visible Whether the dialog is visible to the user.
  * If `false`:
  * - internal state of [DialogWindow] is preserved and will be restored next time the dialog
  * will be visible;
  * - native resources will not be released. They will be released only when [DialogWindow]
  * will leave the composition.
- * @param title Title in the titlebar of the dialog
- * @param icon Icon in the titlebar of the window (for platforms which support this).
+ * @param title Title in the title bar of the dialog
+ * @param icon Icon in the title bar of the window (for platforms that support this).
  * On macOs individual windows can't have a separate icon. To change the icon in the Dock,
  * set it via `iconFile` in build.gradle
  * (https://github.com/JetBrains/compose-jb/tree/master/tutorials/Native_distributions_and_local_execution#platform-specific-options)
  * @param decoration Specifies the decoration for this dialog.
- * @param transparent Disables or enables window transparency. Transparency should be set
- * only if window is undecorated, otherwise an exception will be thrown.
- * @param resizable Can dialog be resized by the user (application still can resize the dialog
- * changing [state])
- * @param enabled Can dialog react to input events
- * @param focusable Can dialog receive focus
- * @param alwaysOnTop Should the dialog always be on top of another windows and dialogs
+ * @param transparent Disables or enables window transparency. Transparency may be set only if the
+ * dialog is undecorated, otherwise an exception will be thrown.
+ * @param resizable Whether the dialog can be resized by the user (application still can resize the
+ * dialog by changing [state]).
+ * @param enabled Whether the dialog reacts to input events.
+ * @param focusable Whether the dialog can receive focus.
+ * @param alwaysOnTop whether the dialog will always be on top of other windows and dialogs in the
+ * application.
  * @param onPreviewKeyEvent This callback is invoked when the user interacts with the hardware
  * keyboard. It gives ancestors of a focused component the chance to intercept a [KeyEvent].
  * Return true to stop propagation of this event. If you return false, the key event will be
  * sent to this [onPreviewKeyEvent]'s child. If none of the children consume the event,
- * it will be sent back up to the root using the onKeyEvent callback.
+ * it will be sent back up to the root using the [onKeyEvent] callback.
  * @param onKeyEvent This callback is invoked when the user interacts with the hardware
  * keyboard. While implementing this callback, return true to stop propagation of this event.
  * If you return false, the key event will be sent to this [onKeyEvent]'s parent.
- * @param content content of the dialog
+ * @param content Composable content of the dialog.
  */
 @ExperimentalComposeUiApi
 @Composable
@@ -397,7 +397,10 @@ fun DialogWindow(
 
 @Deprecated(
     message = "Replaced by SwingDialog",
-    replaceWith = ReplaceWith("SwingDialog(visible, onPreviewKeyEvent, onKeyEvent, create, dispose, update, content)")
+    replaceWith = ReplaceWith(
+        "SwingDialog(visible, onPreviewKeyEvent, onKeyEvent, create, dispose, update, content)",
+        "androidx.compose.ui.awt.SwingDialog"
+    )
 )
 @Composable
 fun Dialog(
@@ -418,19 +421,16 @@ fun Dialog(
     content
 )
 
-// TODO(demin): fix mouse hover after opening a dialog.
-//  When we open a modal dialog, ComposeLayer/mouseExited will
-//  never be called for the parent window. See ./gradlew run3
 /**
  * Compose [ComposeDialog] obtained from [create]. The [create] block will be called
  * exactly once to obtain the [ComposeDialog] to be composed, and it is also guaranteed to
  * be invoked on the UI thread (Event Dispatch Thread).
  *
- * Once Dialog leaves the composition, [dispose] will be called to free resources that
+ * Once [DialogWindow] leaves the composition, [dispose] will be called to free resources that were
  * obtained by the [ComposeDialog].
  *
- * Dialog is a modal window. It means it blocks the parent [Window] / [DialogWindow] in which composition
- * context it was created.
+ * Dialog is a modal window. It means it blocks the parent [Window] / [DialogWindow] in whose
+ * composition context it was created.
  *
  * The [update] block can be run multiple times (on the UI thread as well) due to recomposition,
  * and it is the right place to set [ComposeDialog] properties depending on state.
@@ -440,7 +440,7 @@ fun Dialog(
  * Dialog is needed for creating dialog's that still can't be created with
  * the default Compose function [androidx.compose.ui.window.DialogWindow]
  *
- * @param visible Is [ComposeDialog] visible to user.
+ * @param visible Whether the dialog is visible to the user.
  * If `false`:
  * - internal state of [ComposeDialog] is preserved and will be restored next time the dialog
  * will be visible;
@@ -450,22 +450,22 @@ fun Dialog(
  * keyboard. It gives ancestors of a focused component the chance to intercept a [KeyEvent].
  * Return true to stop propagation of this event. If you return false, the key event will be
  * sent to this [onPreviewKeyEvent]'s child. If none of the children consume the event,
- * it will be sent back up to the root using the onKeyEvent callback.
+ * it will be sent back up to the root using the [onKeyEvent] callback.
  * @param onKeyEvent This callback is invoked when the user interacts with the hardware
  * keyboard. While implementing this callback, return true to stop propagation of this event.
  * If you return false, the key event will be sent to this [onKeyEvent]'s parent.
  * @param create The block creating the [ComposeDialog] to be composed.
  * @param dispose The block to dispose [ComposeDialog] and free native resources.
  * Usually it is simple `ComposeDialog::dispose`
- * @param update The callback to be invoked after the layout is inflated.
- * @param content Composable content of the creating dialog.
+ * @param update The callback to be invoked to update window properties.
+ * @param content Composable content of the dialog.
  */
 @Suppress("unused")
 @Deprecated(
     message = "Renamed to SwingDialog",
     replaceWith = ReplaceWith(
         expression = "SwingDialog(visible, onPreviewKeyEvent, onKeyEvent, create, dispose, update, content)",
-        "androidx.compose.ui.window.SwingDialog"
+        "androidx.compose.ui.awt.SwingDialog"
     )
 )
 @Composable
